GhcPlugins

Documentation

module Plugins

data DynFlags

Constructors

DynFlags_

type FamInstEnvs = (FamInstEnv, FamInstEnv)

data Name

Constructors

Name_

data OccName

Constructors

OccName_

data Module

Constructors

Module_

data UnitId

Constructors

UnitId_

data TyThing

Constructors

TyThing_

data ModSummary

Constructors

ModSummary_

data TcGblEnv

Constructors

TcGblEnv_

data LHsExpr a

Constructors

LHsExpr_

data GhcTc

Constructors

GhcTc_

data GhcRn

Constructors

GhcRn_

data GhcException

Constructors

CmdLineError String

ProgramError String

PprProgramError String SDoc

data ModuleName

Constructors

ModuleName_

data SDoc

Constructors

SDoc_

data HsParsedModule

Constructors

HsParsedModule_

data HsGroup a

Constructors

HsGroup_

data Phase

Constructors

Phase_

data Coercion

Constructors

Coercion_

data Type

Constructors

Type_

type Kind = Type

type TyVar = Var

data TyCoBinder

Constructors

TyCoBinder_

data SrcSpan

Constructors

SrcSpan_

data RealSrcSpan

Constructors

RealSrcSpan_

data Tickish a

Constructors

SourceNote

Fields

sourceSpan :: RealSrcSpan

Source covered

sourceName :: String

Name for source location

data Var

Constructors

Var_

type Id = Var

data Fingerprint

Constructors

Fingerprint_

data PrintUnqualified

Constructors

PrintUnqualified_

data TyCon

Constructors

TyCon_

data IdDetails

Constructors

ClassOpId Class

DataConWorkId DataCon

PrimOpId PrimOp

data IdUnfolding

Constructors

IdUnfolding_

data Unfolding

Constructors

Unfolding_

data FunctionOrData

Constructors

IsFunction

IsData

data FV

Constructors

FV_

data Class

Constructors

Class_

data NameSpace

Constructors

NameSpace_

data HscEnv

Constructors

HscEnv

Fields

hsc_dflags :: DynFlags

data RdrName

Constructors

RdrName_

data Messages

Constructors

Messages_

data Literal

Constructors

LitChar Char
Char# - at least 31 bits. Create with mkMachChar

LitNumber !LitNumType !Integer Type

LitString ByteString
A string-literal: stored and emitted UTF-8 encoded, we'll arrange to decode it at runtime. Also emitted with a '0' terminator. Create with mkMachString

LitNullAddr
The NULL pointer, the only pointer value that can be represented as a Literal. Create with nullAddrLit

LitRubbish
A nonsense value, used when an unlifted binding is absent and has type forall (a :: TYPE UnliftedRep). a. May be lowered by code-gen to any possible value. Also see GHC:Note [Rubbish literals]

LitFloat Rational
Float#. Create with mkMachFloat

LitDouble Rational
Double#. Create with mkMachDouble

LitLabel FastString (Maybe Int) FunctionOrData
A label literal. Parameters:

1) The name of the symbol mentioned in the declaration

2) The size (in bytes) of the arguments the label expects. Only applicable with stdcall labels. Just x => <x> will be appended to label name when emitting assembly.

data LitNumType

Numeric literal type

Constructors

LitNumInteger
Integer (see GHC:Note [Integer literals])

LitNumNatural
Natural (see GHC:Note [Natural literals])

LitNumInt
Int# - according to target machine

LitNumInt64
Int64# - exactly 64 bits

LitNumWord
Word# - according to target machine

LitNumWord64
Word64# - exactly 64 bits

type FastString = String

data AlgTyConRhs

Constructors

AbstractTyCon

DataTyCon

Fields

data_cons :: [DataCon]

TupleTyCon

Fields

data_con :: DataCon

SumTyCon

Fields

data_cons :: [DataCon]

NewTyCon

Fields

data_con :: DataCon

data DataCon

Constructors

DataCon_

data Role

Constructors

Representational

data CoAxiom (a :: BranchFlag)

Constructors

CoAxiom_

data BranchFlag

Constructors

BFBranched

BFUnbranched

type Unbranched = 'BFUnbranched

data PrimOp

Constructors

IntAddOp

IntSubOp

IntMulOp

IntQuotOp

IntRemOp

IntGtOp

IntGeOp

IntLtOp

IntLeOp

IntEqOp

data Expr b

Constructors

Var Id

Lit Literal

App (Expr b) (Arg b)

Lam b (Expr b)

Let (Bind b) (Expr b)

Case (Expr b) b Type [Alt b]

Cast (Expr b) Coercion

Tick (Tickish Id) (Expr b)

Type Type

Coercion Coercion

data Bind b

Constructors

NonRec b (Expr b)

Rec [(b, Expr b)]

data AltCon

Constructors

DataAlt DataCon

LitAlt Literal
A literal: case e of { 1 -> ... } Invariant: always an *unlifted* literal See GHC:Note [Literal alternatives]

DEFAULT
Trivial alternative: case e of { _ -> ... }

type CoreExpr = Expr CoreBndr

type CoreBndr = Var

type CoreAlt = Alt CoreBndr

type CoreBind = Bind CoreBndr

type CoreProgram = [CoreBind]

type Arg b = Expr b

type Alt b = (AltCon, [b], Expr b)

data ImpDeclSpec

Constructors

ImpDeclSpec

Fields

is_mod :: ModuleName

Module imported, e.g. import Muggle Note the Muggle may well not be the defining module for this thing!

is_as :: ModuleName

Import alias, e.g. from as M (or Muggle if there is no as clause)

is_qual :: Bool

Was this import qualified?

is_dloc :: SrcSpan

The location of the entire import declaration

data ImpItemSpec

Constructors

ImpAll

ImpSome

Fields

is_explicit :: Bool

is_iloc :: SrcSpan

data ImportSpec

Constructors

ImpSpec

Fields

is_decl :: ImpDeclSpec

is_item :: ImpItemSpec

data ModIface

Constructors

ModIface

Fields

mi_exports :: [IfaceExport]

data ModGuts

Constructors

ModGuts

Fields

mg_fam_inst_env :: FamInstEnv

mg_module :: Module

data FamInstEnv

Constructors

FamInstEnv_

data SimplMode

Constructors

SimplMode

Fields

sm_names :: [String]

sm_phase :: CompilerPhase

sm_dflags :: DynFlags

sm_rules :: Bool

sm_inline :: Bool

sm_case_case :: Bool

sm_eta_expand :: Bool

data CompilerPhase

Constructors

Phase PhaseNum

InitialPhase

type PhaseNum = Int

data CoreToDo

Constructors

CoreDoSimplify Int SimplMode

CoreDoPluginPass String CorePluginPass

CoreDoFloatInwards

CoreDoFloatOutwards FloatOutSwitches

CoreLiberateCase

CoreDoPrintCore

CoreDoStaticArgs

CoreDoCallArity

CoreDoExitify

CoreDoStrictness

CoreDoWorkerWrapper

CoreDoSpecialising

CoreDoSpecConstr

CoreCSE

CoreDoRuleCheck CompilerPhase String

CoreDoNothing

CoreDoPasses [CoreToDo]

CoreDesugar

CoreDesugarOpt

CoreTidy

CorePrep

CoreOccurAnal

data GlobalRdrElt

Constructors

GRE

Fields

gre_name :: Name

gre_par :: Parent

gre_lcl :: Bool

True = the thing was defined locally

gre_imp :: [ImportSpec]

In scope through these imports

data Parent

Constructors

Parent_

data FloatOutSwitches

Constructors

FloatOutSwitches_

type CorePluginPass = ModGuts -> CoreM ModGuts

type IfaceExport = AvailInfo

data AvailInfo

Constructors

AvailInfo_

data GlobalRdrEnv

Constructors

GlobalRdrEnv_

data UniqSet a

Constructors

UniqSet_

data StubM a

Constructors

StubM_

newtype Hsc a

Constructors

Hsc_ (StubM a)

newtype CoreM a

Constructors

CoreM_ (StubM a)

newtype TcM a

Constructors

TcM_ (StubM a)

type TcRn a = TcM a

newtype IfM ab a

Constructors

IfM_ (StubM a)

newtype Ghc a

Constructors

Ghc_ (StubM a)

data ModLocation

Constructors

ModLocation_

data ModuleOrigin

Constructors

ModuleOrigin_

data UnusablePackageReason

Constructors

UnusablePackageReason_

data ModuleSuggestion

Constructors

ModuleSuggestion_

data FindResult

Constructors

Found ModLocation Module
The module was found

NoPackage UnitId
The requested package was not found

FoundMultiple [(Module, ModuleOrigin)]
_Error_: both in multiple packages

NotFound
Not found

Fields

fr_paths :: [FilePath]

fr_pkg :: Maybe UnitId

fr_mods_hidden :: [UnitId]

fr_pkgs_hidden :: [UnitId]

fr_unusables :: [(UnitId, UnusablePackageReason)]

fr_suggestions :: [ModuleSuggestion]

class Outputable a where

Minimal complete definition

Nothing

Methods

ppr :: a -> SDoc

pprPrec :: Rational -> a -> SDoc

occName :: Name -> OccName

occNameString :: OccName -> String

moduleName :: Module -> ModuleName

nameModule_maybe :: Name -> Maybe Module

moduleUnitId :: Module -> UnitId

stableModuleNameCmp :: ModuleName -> ModuleName -> Ordering

stableUnitIdCmp :: UnitId -> UnitId -> Ordering

fingerprintString :: String -> Fingerprint

fingerprintFingerprints :: [Fingerprint] -> Fingerprint

(<+>) :: SDoc -> SDoc -> SDoc

text :: String -> SDoc

mi_module :: ModIface -> Module

showSDocForUser :: DynFlags -> PrintUnqualified -> SDoc -> String

alwaysQualify :: PrintUnqualified

tyConsOfType :: Type -> UniqSet TyCon

mkCoercionTy :: Coercion -> Type

varType :: Var -> Type

isLiftedTypeKind :: Kind -> Bool

classifiesTypeWithValues :: Kind -> Bool

isRuntimeRepTy :: Type -> Bool

isRuntimeRepKindedTy :: Type -> Bool

splitFunTy_maybe :: Type -> Maybe (Type, Type)

unitTy :: Type

unitDataConId :: Id

voidPrimId :: Id

voidArgId :: Id

rUNTIME_ERROR_ID :: Id

getOccString :: a -> String

getName :: a -> Name

tyVarKind :: a -> Kind

tyConKind :: a -> Kind

tyThingId :: a -> Id

tyThingTyCon :: a -> TyCon

boolTy :: Type

stringTy :: Type

charTy :: Type

getOccName :: a -> OccName

trueDataCon :: DataCon

falseDataCon :: DataCon

unitDataCon :: DataCon

charDataCon :: DataCon

boolTyCon :: TyCon

listTyCon :: TyCon

intTyCon :: TyCon

intPrimTyCon :: TyCon

addrPrimTyCon :: TyCon

voidPrimTyCon :: TyCon

unitTyCon :: TyCon

nonDetEltsUniqSet :: UniqSet a -> [a]

normaliseType :: a -> b -> c -> (Coercion, Type)

getTyVar_maybe :: a -> Maybe TyVar

splitTyConApp_maybe :: Type -> Maybe (TyCon, [Type])

splitAppTy_maybe :: Type -> Maybe (Type, Type)

splitForAllTy_maybe :: Type -> Maybe (TyVar, Type)

splitCastTy_maybe :: Type -> Maybe (Type, Coercion)

splitPiTy_maybe :: Type -> Maybe (TyCoBinder, Type)

unwrapNewTyCon_maybe :: TyCon -> Maybe ([TyVar], Type, CoAxiom Unbranched)

tyConTyVars :: TyCon -> [TyVar]

unionManyUniqSets :: [UniqSet a] -> UniqSet a

dataConTyCon :: DataCon -> TyCon

dataConOrigArgTys :: DataCon -> [Type]

dataConInstOrigArgTys :: DataCon -> [Type] -> [Type]

dataConOrigResTy :: DataCon -> Type

isAlgTyCon :: TyCon -> Bool

isTupleTyCon :: TyCon -> Bool

isFamilyTyCon :: TyCon -> Bool

isStrLitTy :: Type -> Maybe FastString

($+$) :: a -> b -> b infixl 5

algTyConRhs :: TyCon -> AlgTyConRhs

mkCharExpr :: Char -> CoreExpr

isDefaultAlt :: a -> Bool

mkListExpr :: Type -> [CoreExpr] -> CoreExpr

findAlt :: AltCon -> [(AltCon, a, b)] -> Maybe (AltCon, a, b)

errorIds :: [Id]

fvVarList :: FV -> [Var]

isTyVar :: a -> Bool

unpackCStringName :: Name

unpackCStringFoldrName :: Name

buildName :: Name

idDetails :: Id -> IdDetails

realIdUnfolding :: Id -> Unfolding

mkDictSelRhs :: Class -> Int -> CoreExpr

classAllSelIds :: Class -> [Id]

expr_fvs :: CoreExpr -> FV

maybeUnfoldingTemplate :: Unfolding -> Maybe CoreExpr

mkImpossibleExpr :: Type -> CoreExpr

getDynFlags :: Monad m => m DynFlags

mkFastString :: String -> FastString

moduleNameString :: ModuleName -> String

findExposedPackageModule :: HscEnv -> ModuleName -> Maybe FastString -> IO FindResult

mkModule :: UnitId -> ModuleName -> Module

initTcInteractive :: HscEnv -> TcM a -> IO (Messages, Maybe a)

initIfaceTcRn :: a -> TcRn b

mkTyConApp :: TyCon -> [Type] -> Type

lookupId :: Monad m => Name -> m Id

lookupTyCon :: Monad m => Name -> m TyCon

mkRuntimeErrorApp :: Id -> Type -> String -> CoreExpr

idName :: Id -> Name

showPpr :: DynFlags -> a -> String

lookupThing :: Monad m => Name -> m TyThing

mkTyConTy :: TyCon -> Type

mkCoreApps :: CoreExpr -> [CoreExpr] -> CoreExpr

mkIntExpr :: DynFlags -> Integer -> CoreExpr

mkModuleName :: String -> ModuleName

noSrcSpan :: SrcSpan

throwGhcExceptionIO :: GhcException -> IO a

showSDoc :: DynFlags -> SDoc -> String

cannotFindModule :: DynFlags -> ModuleName -> FindResult -> SDoc

hsep :: [SDoc] -> SDoc

panic :: String -> a

lookupGRE_RdrName :: RdrName -> GlobalRdrEnv -> [GlobalRdrElt]

gresFromAvails :: Maybe ImportSpec -> [AvailInfo] -> [GlobalRdrElt]

mkGlobalRdrEnv :: [GlobalRdrElt] -> GlobalRdrEnv

loadPluginInterface :: SDoc -> Module -> IfM lcl ModIface

getHscEnv :: Monad m => m HscEnv

type PackageFamInstEnv = FamInstEnv

getPackageFamInstEnv :: CoreM PackageFamInstEnv

mkUnqual :: NameSpace -> FastString -> RdrName

bindsOnlyPass :: (CoreProgram -> CoreM CoreProgram) -> ModGuts -> CoreM ModGuts

varName :: NameSpace

dataName :: NameSpace

tcClsName :: NameSpace

tyConAppTyCon_maybe :: Type -> Maybe TyCon

nameOccName :: Name -> OccName

charTyConName :: Name

noinlineIdName :: Name

nilDataCon :: DataCon

dataConWorkId :: DataCon -> Id

thNameToGhcName :: Name -> CoreM (Maybe Name)

showSDocUnsafe :: SDoc -> String

dropRuntimeRepArgs :: [Type] -> [Type]